+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [10.0.0] - 2021/05/15
+
+* Breaking change: The new `UserExceptionDialog.useLocalContext` parameter now allows
+  the `UserExceptionDialog` to be put in the `builder` parameter of the `MaterialApp` widget. Even
+  if you use this dialog, it is unlikely this will be a breaking change for you. But if it is (your
+  error dialog now has problems), simply make `useLocalContext: true` to return to the old behavior.
+
 # [9.0.9] - 2021/05/10
 
 * LocalPersist `saveJson()` and `loadJson()` methods.